% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_fit_Bayes.R
\name{DM_fit_Bayes}
\alias{DM_fit_Bayes}
\title{Estimate beta coefficients of the Dirichlet-Multinomial regression model
using Bayes approach with a flat prior.}
\usage{
DM_fit_Bayes(Y, X, iterations, chains)
}
\arguments{
\item{Y}{Matrix of \verb{n x J} responses.}

\item{X}{Matrix of \verb{n x (P+1)} covariates.}

\item{iterations}{Number of iterations for MCMC posterior sampeling.}

\item{chains}{Number of chains for MCMC posterior sampeling.}
}
\value{
point estimation of \verb{(P+1) x J} beta coefficients.

Samples via MCMC.
}
\description{
Estimate beta coefficients of the Dirichlet-Multinomial regression model
using Bayes approach with a flat prior.
}
